CSM matrix fix in previous commit, cleanup math stuff in files and filters
