+++ source/Engine/ETEngine.vcxproj.filters
@@ -165,9 +165,6 @@
     <ClCompile Include="Helper\BinaryReader.cpp">
       <Filter>Helper</Filter>
     </ClCompile>
-    <ClCompile Include="Helper\MathHelper.cpp">
-      <Filter>Helper</Filter>
-    </ClCompile>
     <ClCompile Include="Helper\PerformanceInfo.cpp">
       <Filter>Helper</Filter>
     </ClCompile>
@@ -208,6 +205,7 @@
       <Filter>StaticDependancies\glad</Filter>
     </ClCompile>
     <ClCompile Include="stdafx.cpp" />
+    <ClCompile Include="Math\Geometry.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Base\Context.hpp">
@@ -259,9 +257,6 @@
       <Filter>Content</Filter>
     </ClInclude>
     <ClInclude Include="Helper\Logger.hpp">
-      <Filter>Helper</Filter>
-    </ClInclude>
-    <ClInclude Include="Helper\MathHelper.hpp">
       <Filter>Helper</Filter>
     </ClInclude>
     <ClInclude Include="Helper\Singleton.hpp">
@@ -428,5 +423,7 @@
     </ClInclude>
     <ClInclude Include="stdafx.hpp" />
     <ClInclude Include="Graphics\Frustum.hpp" />
+    <ClInclude Include="Math\MathUtil.hpp" />
+    <ClInclude Include="Math\Geometry.hpp" />
   </ItemGroup>
 </Project>
