+++ source/Engine/Graphics/Frustum.hpp
@@ -1,7 +1,5 @@
 #pragma once
 
-#include "../Helper/MathHelper.hpp"
-
 class CameraComponent;
 
 enum class VolumeCheck
